Read transactions from Google Sheets
Updated README and python file to take TX data from the Google sheet and create a list of transactions and fees.  Airdops must be added manually.  Gas fees are first declared as gains (or losses) and then are declared as fees.  If a sale is closed in losses, then purchases within two months that are not sold do not count towards the losses and are subtracted out.